.yoyo/snapshot/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
boto3>=1.35.77
requests-oauthlib>=2.0.0
cryptography>=43.0.3
python-dotenv>=1.0.1
# Motor 3.6.0 requires pymongo < 4.10; align accordingly
pymongo>=4.9.0,<4.10
pydantic>=2.10.3
pydantic-settings>=2.7.0
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.2.1
argon2-cffi>=23.1.0
passlib>=1.7.4
tzdata>=2024.2
motor==3.6.0
pytest>=8.3.4
pytest-asyncio>=0.24.0
httpx>=0.28.1
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.13.0
requests>=2.32.3
orjson>=3.10.12
pandas>=2.2.3
numpy>=2.2.1
python-multipart>=0.0.17
jq>=1.8.0
typer>=0.15.1
pyodbc>=5.2.0
openpyxl>=3.1.5
Pillow>=11.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.26.0
redis>=5.2.1
psutil>=6.1.0
authlib>=1.4.0
xlsxwriter>=3.2.0

# NOTE: JWT usage has been migrated to an Authlib-based compatibility layer
# (backend/auth/jwt_provider.py) to avoid transitive ECDSA dependencies.
# Removed `python-jose` from requirements intentionally. If you need
# asymmetric JWT algorithms (ES/RSA), add a vetted library and load keys
# from secure storage. See REMEDIATION_PLAN.md for details.

# Security/upstream fixes (conservative bumps applied 2025-11-06)
# Upgrade setuptools to address known advisories
setuptools>=78.1.1
aiohttp>=3.9.0
memory-profiler>=0.61.0
asgi-lifespan>=2.1.0
